requirements.txt
fastapi 
uvicorn 
openai 
numpy 
scikit-learn 
python-dotenv